+++ Pocket Casts TV App/UI/Player/NowPlayingViewModel.swift
@@ -29,6 +29,7 @@ class NowPlayingViewModel: Identifiable {
     var isLoading: Bool = true
     var isFirstLoad: Bool = false
     var isFailed: Bool = false
+    var isPlaying: Bool = false
 
     @ObservationIgnored private var timeControlStatusObservation: NSKeyValueObservation?
     @ObservationIgnored private var itemStatusObservation: NSKeyValueObservation?
@@ -115,6 +116,7 @@ class NowPlayingViewModel: Identifiable {
         let itemNotReady = status == .unknown
         isLoading = waiting || itemNotReady
         isFailed = status == .failed
+        isPlaying = player.timeControlStatus == .playing
         if !isLoading {
             isFirstLoad = false
         }